requirements.txt
PyQt6>=6.5.0
Pillow>=9.0.0
PyYAML>=6.0
platformdirs>=4.2.0